CloudSage-main/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
google-generativeai>=0.7.2
python-multipart>=0.0.9
pydantic>=2.10.0
python-dotenv>=1.0.1
pandas>=2.2.3
